+++ packages/library-base/src/commonMain/kotlin/io/realm/schema/RealmStorageType.kt
@@ -18,6 +18,7 @@ package io.realm.schema
 
 import io.realm.BaseRealmObject
 import io.realm.EmbeddedObject
+import io.realm.ObjectId
 import io.realm.RealmInstant
 import io.realm.RealmObject
 import kotlin.reflect.KClass
@@ -61,5 +62,10 @@ public enum class RealmStorageType(public val kClass: KClass<*>) {
     /**
      * Storage type for properties of type [RealmInstant].
      */
-    TIMESTAMP(RealmInstant::class);
+    TIMESTAMP(RealmInstant::class),
+
+    /**
+     * Storage type for properties of type [ObjectId].
+     */
+    OBJECT_ID(ObjectId::class);
 }