+++ packages/plugin-compiler/src/test/resources/sample/input/Sample.kt
@@ -17,6 +17,7 @@
 package sample.input
 
 import io.realm.EmbeddedObject
+import io.realm.ObjectId
 import io.realm.RealmInstant
 import io.realm.RealmList
 import io.realm.RealmObject
@@ -50,6 +51,7 @@ class Sample : RealmObject {
     var floatField: Float? = 3.14f
     var doubleField: Double? = 1.19840122
     var timestampField: RealmInstant? = RealmInstant.fromEpochSeconds(0,0)
+    var objectIdField: ObjectId? = ObjectId.create()
     var child: Child? = null
 
     // List types
@@ -63,6 +65,7 @@ class Sample : RealmObject {
     var floatListField: RealmList<Float> = realmListOf()
     var doubleListField: RealmList<Double> = realmListOf()
     var timestampListField: RealmList<RealmInstant> = realmListOf()
+    var objectIdListField: RealmList<ObjectId> = realmListOf()
     var objectListField: RealmList<Sample> = realmListOf()
     var embeddedObjectListField: RealmList<EmbeddedChild> = realmListOf()
 
@@ -77,6 +80,7 @@ class Sample : RealmObject {
     var nullableFloatListField: RealmList<Float?> = realmListOf()
     var nullableDoubleListField: RealmList<Double?> = realmListOf()
     var nullableTimestampListField: RealmList<RealmInstant?> = realmListOf()
+    var nullableObjectIdListField: RealmList<ObjectId?> = realmListOf()
 
     fun dumpSchema(): String = "${Sample.`io_realm_kotlin_schema`()}"
 }